+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
-import List from "./List.jsx";
-import Author from "./Author.jsx";
+import List from "./component/list/List.jsx";
+import Author from "./component/author/Author.jsx";
 import Header from './header.js';
+// import { BrowserRouter as Router } from 'react-router-dom';
 import UserPage from './component/userPage/userPage.js';
